+++ prefs_sembast/test/db_sembast_test.dart
@@ -4,10 +4,12 @@ import 'package:sembast/sembast_memory.dart';
 import 'package:tekartik_prefs_sembast/prefs.dart';
 import 'package:tekartik_prefs_sembast/prefs_async.dart';
 import 'package:tekartik_prefs_sembast/prefs_light.dart';
+import 'package:tekartik_prefs_test/kv_store_test_runner.dart';
 import 'package:tekartik_prefs_test/prefs_async_test_runner.dart';
 import 'package:tekartik_prefs_test/prefs_async_with_cache_test_runner.dart';
 import 'package:tekartik_prefs_test/prefs_light_test_runner.dart';
 import 'package:tekartik_prefs_test/prefs_test_runner.dart' as prefs;
+import 'package:test/test.dart';
 
 void main() {
   runPrefsAsyncTests(
@@ -23,4 +25,12 @@ void main() {
       path: '.',
     ),
   );
+  group('kv_store', () {
+    runKvStoreTests(
+      getPrefsLightSembast(
+        databaseFactory: newDatabaseFactoryMemory(),
+        path: '.',
+      ),
+    );
+  });
 }